Stability improvements in test suites.

+++ util/product.py
@@ -6,6 +6,7 @@ import json
 import logging
 import os
 from rpc.rpc_call import RPC
+import shutil
 import subprocess
 import time
 
@@ -22,10 +23,10 @@ class Product():
    _logs = []
    _processes = []
    _userProductConfig = None
-   _resultsDir = None
+   _cmdlineArgs = None
 
-   def __init__(self, resultsDir, apiServerUrl, userProductConfig):
-      self._resultsDir = resultsDir
+   def __init__(self, cmdlineArgs, apiServerUrl, userProductConfig):
+      self._cmdlineArgs = cmdlineArgs
       self._apiServerUrl = apiServerUrl
       self._userProductConfig = userProductConfig
 
@@ -35,7 +36,7 @@ class Product():
       Raises an exception if it cannot start.
       '''
       atexit.register(self.stopProduct)
-      productLogsDir = os.path.join(self._resultsDir, PRODUCT_LOGS_DIR)
+      productLogsDir = os.path.join(self._cmdlineArgs.resultsDir, PRODUCT_LOGS_DIR)
       os.makedirs(productLogsDir, exist_ok=True)
 
       # since we change directories while launching products, save cwd here
@@ -47,6 +48,10 @@ class Product():
             projectSection = launchElement[project]
             buildRoot = projectSection["buildRoot"]
             buildRoot = os.path.expanduser(buildRoot)
+
+            if not self._cmdlineArgs.keepAthenaDB and \
+               project.lower().startswith("athena"):
+               self.clearAthenaDB(launchElement[project])
 
             for executable in projectSection:
                if executable == "buildRoot":
@@ -62,7 +67,7 @@ class Product():
                   previousParam = None
                   for param in executableSection["parameters"]:
                      if executable.startswith("replica") and previousParam == "-d":
-                        param = os.path.join(self._resultsDir, param)
+                        param = os.path.join(self._cmdlineArgs.resultsDir, param)
                         os.makedirs(param)
 
                      cmd.append(os.path.expanduser(param))
@@ -82,14 +87,56 @@ class Product():
       if not self._waitForProductStartup():
          raise Exception("The product did not start. Exiting.")
 
+   def clearAthenaDB(self, athenaSection):
+      '''
+      Deletes the Athena DB so we get a clean start.
+      Other test suites can leave it in a state that makes
+      it fail.
+      '''
+      params = None
+
+      for subSection in athenaSection:
+         if subSection.lower().startswith("athena"):
+            params = athenaSection[subSection]["parameters"]
+
+      isConfigParam = False
+
+      for param in params:
+         if isConfigParam:
+            configFile = os.path.join(athenaSection["buildRoot"],
+                                      param)
+            configFile = os.path.expanduser(configFile)
+            subPath = None
+
+            with open (configFile, "r") as props:
+               for prop in props:
+                  prop = prop.strip()
+                  if prop and not prop.startswith("#") and \
+                     prop.lower().startswith("blockchain_db_path"):
+                     subPath = prop.split("=")[1]
+
+            dbPath = os.path.join(athenaSection["buildRoot"], subPath)
+            dbPath = os.path.expanduser(dbPath)
+            if os.path.isdir(dbPath):
+               log.debug("Clearing Athena DB directory '{}'".format(dbPath))
+               shutil.rmtree(dbPath)
+
+         if param == "-c":
+            isConfigParam = True
 
    def stopProduct(self):
       '''
       Stops the product executables and closes the logs.
       '''
       for p in self._processes[:]:
-         p.terminate()
-         self._processes.remove(p)
+         if p.poll() is None:
+            p.terminate()
+            print("Terminating {}.".format(p.args))
+
+      for p in self._processes[:]:
+         while p.poll() is None:
+            print("Waiting for process {} to exit.".format(p.args))
+            time.sleep(1)
 
       for log in self._logs[:]:
          log.close()
@@ -117,7 +164,7 @@ class Product():
       attempts = 0
       # Helen now takes ~7-8 seconds to boot so we should wait for around 10 seconds
       sleepTime = 10
-      startupLogDir = os.path.join(self._resultsDir, PRODUCT_LOGS_DIR,
+      startupLogDir = os.path.join(self._cmdlineArgs.resultsDir, PRODUCT_LOGS_DIR,
                                    "waitForStartup")
       rpc = RPC(startupLogDir, "waitForStartup", self._apiServerUrl)
       caller = self._userProductConfig["users"][0]["hash"]
